requirements.txt
# VoiceToVision - Dependencias

# Bot Discord
py-cord>=2.4.1
discord.py>=2.3.2

# Transcripción de audio
openai-whisper>=20231117

# Procesamiento de audio
ffmpeg-python>=0.2.0
pydub>=0.25.1

# API y requests
requests>=2.31.0
aiohttp>=3.9.1

# Base de datos
aiosqlite>=0.19.0

# Utilidades
python-dotenv>=1.0.0
aiofiles>=23.2.1
cryptography>=41.0.7

# Utilidades de sistema
psutil>=5.9.6
tqdm>=4.66.1

# WhatsApp (opcional - para futura integración)
# twilio>=8.10.0
